Decouple window focus from clicking and add support for window_at (#86)
We introduce a new action focus_at(x,y) to separate window focusing from clicking. This way we can be responsive to differences in the click_offset setting across different apps or contexts.

We also provide a new option to use the beta API window_at. It is currently buggy so it is disabled by default.